# frontend/src/api/generated/current-openapi.yaml
openapi: 3.1.0
info:
  title: ShuttleCube Internal Management API
  version: 0.1.0
  description: |
    羽毛球培训与场地经营管理系统的同源内部接口契约。
    该接口仅供已登录管理界面使用，不是外部公共 API。
    `/agents`、`/workflows`、`/agent-runs` 和 `/approval-requests` 为未来 Agent 契约草案，
    当前版本不得实现、生成客户端调用或作为验收范围。
servers:
  - url: /api/v1
security:
  - sessionCookie: []
...
paths:
...
  /operations/settings/model:
    get:
      tags:
        - Operations Context
      operationId: getOperationsModelSetting
      summary: Return the current Venue model opt-in state
      responses:
        "200":
          description: Model setting; credentials and secrets are never returned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ModelSetting"
        "401":
          $ref: "#/components/responses/Problem"
        "403":
          $ref: "#/components/responses/Problem"
    patch:
      tags:
        - Operations Context
      operationId: updateOperationsModelSetting
      summary: Explicitly enable or disable model use for the current Venue
      parameters:
        - $ref: "#/components/parameters/CsrfToken"
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ModelSettingUpdate"
      responses:
        "200":
          description: Updated model setting
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ModelSetting"
        "401":
          $ref: "#/components/responses/Problem"
        "403":
          $ref: "#/components/responses/Problem"
        "409":
          $ref: "#/components/responses/Problem"
  /operations/settings/runtime:
    get:
      tags:
        - Operations Context
      operationId: getOperationsRuntimeSetting
      summary: Return deterministic operations and write-tool feature flags
      responses:
        "200":
          description: Current Venue runtime setting
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RuntimeSetting"
        "401":
          $ref: "#/components/responses/Problem"
        "403":
          $ref: "#/components/responses/Problem"
    patch:
      tags:
        - Operations Context
      operationId: updateOperationsRuntimeSetting
      summary: Enable or disable deterministic operations and approved write tools
      parameters:
        - $ref: "#/components/parameters/CsrfToken"
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RuntimeSettingUpdate"
      responses:
        "200":
          description: Updated runtime setting
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RuntimeSetting"
        "401":
          $ref: "#/components/responses/Problem"
        "403":
          $ref: "#/components/responses/Problem"
        "409":
          $ref: "#/components/responses/Problem"
        "422":
          $ref: "#/components/responses/Problem"
  /operations/memberships:
    get:
      tags:
        - Operations Context
      operationId: listOperationsMemberships
      summary: List current-Venue memberships for permission review
      responses:
        "200":
          description: Memberships and derived capability bundles
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MembershipSummary"
        "401":
          $ref: "#/components/responses/Problem"
        "403":
          $ref: "#/components/responses/Problem"
  /operations/memberships/{membership_id}:
    patch:
      tags:
        - Operations Context
      operationId: updateOperationsMembership
      summary: Review, enable, disable or change a fixed Venue role bundle
      parameters:
        - $ref: "#/components/parameters/MembershipId"
        - $ref: "#/components/parameters/CsrfToken"
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MembershipUpdate"
      responses:
        "200":
          description: Updated membership; invalid assignees are deterministically requeued
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MembershipSummary"
        "401":
          $ref: "#/components/responses/Problem"
        "403":
          $ref: "#/components/responses/Problem"
        "404":
          $ref: "#/components/responses/Problem"
        "409":
          $ref: "#/components/responses/Problem"
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: shuttlecube_session
  parameters:
...
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema:
        type: string
        minLength: 8
        maxLength: 200
...
    CsrfToken:
      name: X-CSRF-Token
      in: header
      required: true
      schema:
        type: string
        minLength: 16
        maxLength: 300
...
    MembershipId:
      name: membership_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  responses:
...
    Problem:
      description: RFC 9457-style application problem
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
  schemas:
...
    Problem:
      type: object
      additionalProperties: true
      required:
        - type
        - title
        - status
        - detail
        - instance
      properties:
        type:
          type: string
          format: uri
        title:
          type: string
        status:
          type: integer
          minimum: 400
          maximum: 599
        detail:
          type: string
        instance:
          type: string
...
    ModelSetting:
      type: object
      additionalProperties: false
      required:
        - model_enabled
        - provider_configured
        - updated_at
        - version
      properties:
        model_enabled:
          type: boolean
        provider_configured:
          type: boolean
        enabled_by:
          type:
            - string
            - "null"
          format: uuid
        enabled_at:
          type:
            - string
            - "null"
          format: date-time
        updated_at:
          type: string
          format: date-time
        version:
          type: integer
          minimum: 1
    ModelSettingUpdate:
      type: object
      additionalProperties: false
      required:
        - model_enabled
        - reason
        - expected_version
      properties:
        model_enabled:
          type: boolean
        reason:
          type: string
          minLength: 1
          maxLength: 500
        expected_version:
          type: integer
          minimum: 1
    RuntimeSetting:
      type: object
      additionalProperties: false
      required:
        - operations_enabled
        - write_tools_enabled
        - updated_at
        - version
      properties:
        operations_enabled:
          type: boolean
        write_tools_enabled:
          type: boolean
        updated_at:
          type: string
          format: date-time
        version:
          type: integer
          minimum: 1
    RuntimeSettingUpdate:
      type: object
      additionalProperties: false
      required:
        - operations_enabled
        - write_tools_enabled
        - reason
        - expected_version
      properties:
        operations_enabled:
          type: boolean
        write_tools_enabled:
          type: boolean
        reason:
          type: string
          minLength: 1
          maxLength: 500
        expected_version:
          type: integer
          minimum: 1
    MembershipSummary:
      type: object
      additionalProperties: false
      required:
        - id
        - user_id
        - display_name
        - status
        - role_key
        - capabilities
        - version
      properties:
        id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
        display_name:
          type: string
        status:
          type: string
          enum:
            - pending_review
            - active
            - disabled
        role_key:
          type: string
          enum:
            - owner
            - operations_manager
            - operator
            - finance_viewer
        capabilities:
          type: array
          uniqueItems: true
          items:
            type: string
        reviewed_by:
          type:
            - string
            - "null"
          format: uuid
        reviewed_at:
          type:
            - string
            - "null"
          format: date-time
        version:
          type: integer
          minimum: 1
    MembershipUpdate:
      type: object
      additionalProperties: false
      required:
        - status
        - role_key
        - reason
        - expected_version
      properties:
        status:
          type: string
          enum:
            - active
            - disabled
        role_key:
          type: string
          enum:
            - owner
            - operations_manager
            - operator
            - finance_viewer
        reason:
          type: string
          minLength: 1
          maxLength: 500
        expected_version:
          type: integer
          minimum: 1
...